moved userStore to src root
